Increase understandability and Just Works factor
It's confusing to have the plugin be run on the plugin's .cpp file
itself, with all the complicated commandline options needed for even
getting that to compile; this avoids a bit of a "yo dawg" effect. Also,
make some minor tweaks to the Makefile in order to support that change
and to make other things Just Work.